Sidekiq and Puma don't remove their pidfile, do it for them

+++ contrib/scripts/server_puma.sh
@@ -22,21 +22,27 @@
 # redmine$ server_puma.sh start
 # redmine$ server_puma.sh stop
 # redmine$ server_puma.sh restart
-#
 
 SERVER_NAME="redmine"
 
 RAILS_ENV="production"
 
 REDMINE_PATH="$HOME/redmine"
-BIND_URI="unix://$REDMINE_PATH/tmp/sockets/redmine.sock"
-PID_FILE="$REDMINE_PATH/tmp/pids/puma.pid"
 CONFIG_FILE="$HOME/etc/puma.rb"
+
+PID_FILE="$REDMINE_PATH/tmp/pids/puma.pid"
+SOCK_FILE="$REDMINE_PATH/tmp/sockets/redmine.sock"
+
+BIND_URI="unix://$SOCK_FILE"
+
+THREADS="0:8"
+WORKERS=2
 
 function start () {
   echo "Start Puma Server..."
   puma --daemon --preload --bind $BIND_URI \
        --environment $RAILS_ENV --dir $REDMINE_PATH \
+       --workers $WORKERS --threads $THREADS \
        --pidfile $PID_FILE --tag $SERVER_NAME \
        --config $CONFIG_FILE
   echo "Done"
@@ -44,7 +50,11 @@ function start () {
 
 function stop () {
   echo "Stop Puma Server..."
-  kill $(cat $PID_FILE) 2>/dev/null
+  if [ -f $PID_FILE ] ; then
+    kill $(cat $PID_FILE) 2>/dev/null
+    rm -f $PID_FILE
+    rm -f $SOCK_FILE
+  fi
   echo "Done"
 }
